resolve UI-bounded 64-to-32 truncations
Popup tags, stepper values, table rows and collection counts are all
bounded by the UI or by user-maintained lists, so narrowing them is safe.
Where a type was simply wrong, fix the type instead of casting:

  - -numberOfRowsInTableView: now returns NSInteger, as
    NSTableViewDataSource declares it.
  - -errorDidEnd:returnCode: takes NSInteger, matching the alert response
    it receives.
  - ProfileManager's form handler narrows popup tags once where the value
    leaves AppKit rather than at each of fifteen call sites.
  - setIntValue:/numberWithInt: become the NSInteger-taking variants.

+++ cotvnc/Source/ConnectionWaiter.m
@@ -28,7 +28,7 @@
 
 @interface ConnectionWaiter(Private)
 
-- (void)errorDidEnd:(NSWindow *)sheet returnCode:(int)returnCode
+- (void)errorDidEnd:(NSWindow *)sheet returnCode:(NSInteger)returnCode
         contextInfo:(void *)info;
 
 @end
@@ -301,13 +301,13 @@
                 @selector(errorDidEnd:returnCode:contextInfo:), NULL, NULL,
                 @"%@", message);
     else {
-        int ret;
+        NSInteger ret;
         ret = NSRunAlertPanel(theAction, message, ok, NULL, NULL, NULL);
         [self errorDidEnd:nil returnCode:ret contextInfo:nil];
     }
 }
 
-- (void)errorDidEnd:(NSWindow *)sheet returnCode:(int)returnCode
+- (void)errorDidEnd:(NSWindow *)sheet returnCode:(NSInteger)returnCode
         contextInfo:(void *)info
 {
     [delegate connectionFailed];
